pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xero-workflow-automation"
version = "0.2.0"
description = "Xero workflow automation with web UI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.40.0",
    "pydantic>=2.0.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-playwright>=0.4.0",
]

[project.scripts]
xero-app = "app:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "e2e: marks tests as end-to-end tests requiring --live flag",
]
